+++ rules/react.js
@@ -86,6 +86,9 @@ module.exports = {
   // Prevent usage of dangerous JSX properties
   'react/no-danger': 2,
 
+  // Prevent problem with children and props.dangerouslySetInnerHTML
+  'react/no-danger-with-children': 2,
+
   // Prevent usage of deprecated methods
   'react/no-deprecated': 2,