+++ rules/react.js
@@ -122,6 +122,9 @@ module.exports = {
   // Prevent usage of unknown DOM property
   'react/no-unknown-property': 2,
 
+  // Prevent definitions of unused prop types
+  'react/no-unused-prop-types': 2,
+
   // Enforce ES5 or ES6 class for React Components
   'react/prefer-es6-class': 0,