+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add Node v4.5, v6.3, and v6.4 to continuous testing
 
 ### Removed
+- Remove support for Node v5
 - Remove Node v5 from continuous testing
 
 ## [8.0.0] - 2016-08-04