+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All changes to this project are documented in this file. This project adheres to
 - Replace `no-negated-in-lhs` with `no-unsafe-negation`
 - Replace `no-spaced-func` with `func-call-spacing`
 - Move `sort-imports` to ECMAScript 6 rules
+- Silence the log output of `npm run lint`
 
 ### Removed
 - Remove support for Node v5