+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All changes to this project are documented in this file. This project adheres to
 - Replace `no-spaced-func` with `func-call-spacing`
 - Move `sort-imports` to ECMAScript 6 rules
 - Silence the log output of `npm run lint`
+- Exempt `'a'` from `id-length` rule
 
 ### Removed
 - Remove support for Node v5