+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `sort-keys` rule to require object keys to be sorted
 - Add `no-template-curly-in-string` rule to disallow template literal placeholder syntax in regular strings
 - Add a `ryansobol/express` environment
+- Add `props` option to `no-self-assign` rule to disallow self-assignment of properties
 
 ### Changed
 - Replace `no-native-reassign` with `no-global-assign`