+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `props` option to `no-self-assign` rule to disallow self-assignment of properties
 - Add `symbol-description` rule to require symbol description
 - Add `class-methods-use-this` rule to not enforce that class methods utilize `this`
+- Add `react/forbid-component-props` rule to not forbid certain props on Components
 
 ### Changed
 - Update `eslint` dependency to ^3.4.0