+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All changes to this project are documented in this file. This project adheres to
 
 ### Changed
 - Update `eslint` dependency to ^3.4.0
+- Update `eslint-plugin-react` dependency to ^6.2.0
 - Replace `no-native-reassign` with `no-global-assign`
 - Replace `no-negated-in-lhs` with `no-unsafe-negation`
 - Replace `no-spaced-func` with `func-call-spacing`